Backend/requirements.txt
Flask
boto3
openpyxl
python-dotenv
flask-cors
pandas
matplotlib
